Mitigate prototype pollution when filtering options
